Update 3.1-2
The wrapper has been updated to version 3.1-2 with subtle visual improvements and, more importantly, stricter abstraction and checks for the existence of the project tree before starting.
For example, if something is wrong or not found, a warning will be displayed and execution will be interrupted. The process only proceeds if the current checks are successful.
My plans for a future version include adding a pure debugging function or even an executable just for full debugging.
Simply run: `git pull`, inside the cloned directory to update.

+++ install.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
-
-NC="\033[0m"
-GREENCOLOR="\e[1;32m"
-YELLOWCOLOR="\e[1;33m"
-REDCOLOR="\e[1;31m"
+#follow @sakaki91
 
 creationTree(){
     while true; do
@@ -19,7 +15,7 @@ creationTree(){
                 break
             ;;
             *)
-                echo -e "\n[$REDCOLOR ERROR $NC]\n==> Invalid value"
+                echo -e "\n$ERROR_LOG\n==> Invalid value"
                 sleep 2
                 clear
             ;;
@@ -28,22 +24,12 @@ creationTree(){
 }
 
 variableTree(){
-    EXEC_SHORTCUT="$HOME/.local/share/applications/com.sakaki.Teknoparrot.desktop"
+    [ -d "$TREE" ] && rm -rf "$TREE"
     PROGRAM=${TREE}/PROGRAM
     PREFIX=${TREE}/PREFIX
     RUNNER=${TREE}/RUNNER
     TMP=${TREE}/TMP
     RUNNER_EXEC="$RUNNER"/umu-run
-}
-
-fileExistenceChecker(){ 
-    if [[ -d "$TREE" || -f "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop ]]; then
-        rm -rf "$TREE"
-        rm -rf "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    fi
-    if [[ -f "$HOME"/.icons/teknoparrot-icon.png ]]; then
-        rm -r "$HOME"/.icons/teknoparrot-icon.png
-    fi
 }
 
 runner(){
@@ -63,40 +49,46 @@ runner(){
 dependencyInstall(){
     clear
     export WINEPREFIX=${PREFIX}
-    echo -e "[$YELLOWCOLOR WAIT $NC] Proton Wineboot" && $RUNNER_EXEC wineboot -u &> /dev/null
-    if [[ -d $PREFIX/pfx ]]; then
-        echo -e "==> [$GREENCOLOR DONE $NC] Structure created!"
-    else
-        echo -e "==> [$REDCOLOR ERROR $NC] Structure not created!"
-    fi
-    echo -e "[$YELLOWCOLOR WAIT $NC] Downloading .NET Runtime" && wget -c https://aka.ms/dotnet/8.0/dotnet-runtime-win-x64.exe --directory-prefix="$TMP" &> /dev/null
-    echo -e "[$YELLOWCOLOR WAIT $NC] Downloading .NET Desktop Runtime" && wget -c https://aka.ms/dotnet/8.0/windowsdesktop-runtime-win-x64.exe --directory-prefix="$TMP" &> /dev/null
-    echo -e "[$YELLOWCOLOR WAIT $NC] Installing .NET Runtime" && "$RUNNER_EXEC" "$TMP"/dotnet-runtime-win-x64.exe /install /quiet /norestart &> /dev/null && echo -e "==> [$GREENCOLOR DONE $NC] .NET Runtime installed!"
-    echo -e "[$YELLOWCOLOR WAIT $NC] Installing .NET Desktop Runtime" && "$RUNNER_EXEC" "$TMP"/windowsdesktop-runtime-win-x64.exe /install /quiet /norestart &> /dev/null && echo -e "==> [$GREENCOLOR DONE $NC] .NET Desktop Runtime installed!"
-    echo -e "[$YELLOWCOLOR WAIT $NC] Downloading Teknoparrot (Web-Installer)" && wget -c https://github.com/nzgamer41/TPBootstrapper/releases/latest/download/TPBootstrapper.zip --directory-prefix="$TMP" &> /dev/null
+    echo -e "$WAIT_LOG Proton Wineboot." && $RUNNER_EXEC wineboot -u &> /dev/null
+        [ -d "$PREFIX"/pfx ] && echo -e " $DONE_LOG Structure created!"
+        [ ! -d "$PREFIX"/pfx ] && echo -e " $ERROR_LOG Structure not created." && exit
+    echo -e "$WAIT_LOG Downloading .NET Runtime.\n$WAIT_LOG Downloading .NET Desktop Runtime."
+    wget -c https://aka.ms/dotnet/8.0/dotnet-runtime-win-x64.exe --directory-prefix="$TMP" &> /dev/null
+    wget -c https://aka.ms/dotnet/8.0/windowsdesktop-runtime-win-x64.exe --directory-prefix="$TMP" &> /dev/null
+        [ -f "$TMP"/dotnet-runtime-win-x64.exe ] && echo -e " $DONE_LOG .NET Runtime was downloaded!"
+        [ ! -f "$TMP"/dotnet-runtime-win-x64.exe ] && echo -e " $ERROR_LOG .NET Runtime was not downloaded." && exit
+        [ -f "$TMP"/windowsdesktop-runtime-win-x64.exe ] && echo -e " $DONE_LOG .NET Desktop Runtime was downloaded!"
+        [ ! -f "$TMP"/windowsdesktop-runtime-win-x64.exe ] && echo -e " $ERROR_LOG .NET Desktop Runtime was not downloaded." && exit
+        "$RUNNER_EXEC" "$TMP"/dotnet-runtime-win-x64.exe /install /quiet /norestart &> /dev/null && echo -e " $DONE_LOG .NET Runtime installed!"
+        "$RUNNER_EXEC" "$TMP"/windowsdesktop-runtime-win-x64.exe /install /quiet /norestart &> /dev/null && echo -e " $DONE_LOG .NET Desktop Runtime installed!"  
     (
-        echo -e "[$YELLOWCOLOR WAIT $NC] Extracting Teknoparrot (Web-Installer)" && unzip "$TMP"/TPBootstrapper.zip -d "$PROGRAM" &> /dev/null
+        echo -e "$WAIT_LOG Downloading Teknoparrot (Web-Installer)"
+        wget -c https://github.com/nzgamer41/TPBootstrapper/releases/latest/download/TPBootstrapper.zip --directory-prefix="$TMP" &> /dev/null
+            [ -f "$TMP"/TPBootstrapper.zip ] && echo -e " $DONE_LOG TPBootstrapper was downloaded!"
+            [ ! -f "$TMP"/TPBootstrapper.zip ] && echo -e " $ERROR_LOG TPBootstrapper was not downloaded." && exit
+        unzip "$TMP"/TPBootstrapper.zip -d "$PROGRAM" &> /dev/null
+            [ -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $DONE_LOG TPBootstrapper was extracted!"
+            [ ! -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $ERROR_LOG TPBootstrapper was not extracted." && exit
         cd "$PROGRAM"
-        echo -e "[$YELLOWCOLOR WAIT $NC] Installing Teknoparrot (Web-Installer)" && "$RUNNER_EXEC" TPBootstrapper.exe &> /dev/null
-        if [[ -f $PROGRAM/TeknoParrotUi.exe ]]; then
-            echo -e "==> [$GREENCOLOR DONE $NC] Teknoparrot installed!"
-        else
-            echo -e "==> [$REDCOLOR ERROR $NC] Teknoparrot not installed!"
-        fi
+        echo -e "$WAIT_LOG Installing Teknoparrot (Web-Installer)"
+            [ -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $DONE_LOG TPBootstrapper was found!"
+            [ ! -f "$PROGRAM"/TPBootstrapper.exe ] && echo -e " $ERROR_LOG TPBootstrapper was not found." && exit
+            "$RUNNER_EXEC" TPBootstrapper.exe &> /dev/null
+            [ -f "$PROGRAM"/TeknoParrotUi.exe ] && echo -e " $DONE_LOG Teknoparrot installed!"
+            [ ! -f "$PROGRAM"/TeknoParrotUi.exe ] && echo -e " $ERROR_LOG Teknoparrot not installed." && exit
     )
     rm -rf "$PROGRAM"/TPBootstrapper*
-    rm -rf "$TMP"
+    rm -rf "$TMP" && echo -e " $DONE_LOG Temporary files cleared!"
 }
 
 executableCreation(){
+    HEADER="#!/bin/bash"
+    DEBUG_FLAG="#export PROTON_LOG=1"
+    DRIPRIME_FLAG="#export DRI_PRIME=1"
+    MANGOHUD_FLAG="#export MANGOHUD=1"
+    PROTON_LOCATION="export PROTONPATH=$HOME/.local/share/Steam/compatibilitytools.d/UMU-Proton-9.0-4e"
     (
-        mkdir -p "$HOME/.icons"
-        cp -r icon.png "$HOME/.icons/teknoparrot-icon.png"
         cd "$TREE"
-        HEADER="#!/bin/bash"
-        DEBUG_FLAG="#PROTON_LOG=1"
-        DRIPRIME_FLAG="#export DRI_PRIME=1"
-        MANGOHUD_FLAG="#export MANGOHUD=1"
         echo "$HEADER" > Teknoparrot-Linux
         echo "$DEBUG_FLAG" >> Teknoparrot-Linux
         echo "$DRIPRIME_FLAG" >> Teknoparrot-Linux
@@ -104,40 +96,39 @@ executableCreation(){
         echo "export LC_ALL=C" >> Teknoparrot-Linux
         echo "export LC_NUMERIC=C" >> Teknoparrot-Linux
         echo "export LANG=en_US.UTF-8" >> Teknoparrot-Linux
+        echo "export GAMEID=0" >> Teknoparrot-Linux
+        echo "$PROTON_LOCATION" >> Teknoparrot-Linux
         echo "export WINEPREFIX=$PREFIX" >> Teknoparrot-Linux
         echo "$RUNNER_EXEC" "$PROGRAM"/TeknoParrotUi.exe >> Teknoparrot-Linux
         chmod +x Teknoparrot-Linux
     )
-    echo "[Desktop Entry]" > "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    echo "Exec="$TREE"/Teknoparrot-Linux" >> "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    echo "Name=Teknoparrot" >> "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    echo "Icon="$HOME"/.icons/teknoparrot-icon.png" >> "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    echo "Terminal=false" >> "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    echo "Type=Application" >> "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    echo "Categories=Game;" >> "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
-    chmod +x "$HOME"/.local/share/applications/com.sakaki.Teknoparrot.desktop
 }
+
+ARL_NAME="Arcade Runtime Linux"
+ARL_VERSION="3.1-2"
+DONE_LOG="\e[1;32m* [ DONE ]\033[0m"
+WAIT_LOG="\e[1;33m* [ WAIT ]\033[0m"
+ERROR_LOG="\e[1;31m* [ ERROR ]\033[0m"
 
 case $1 in
     "--help")
-        echo -e "\nTeknoparrot-Linux (Runtime): Version 3.1-1\n\n--help\t\tShow this message.\n--debug \tIt executes the debug executable file (this may take some time and usually generates its own log file).\n--remove\tClears all files created by the script.\n"
+        echo -e "\n$ARL_NAME $ARL_VERSION\n\n--help\t\tShow this message.\n--version\tShow wrapper version.\n--debug \tIt executes the debug executable file (this may take some time and usually generates its own log file).\n--remove\tClears all files created by the script.\n"
         exit
     ;;
     "--debug")
         echo -e "\nUNDER DEVELOPMENT\n"
         exit
     ;;
+    "--version")
+        echo -e "$ARL_NAME $ARL_VERSION"
+        exit
+    ;;
     "--remove")
         clear
         creationTree
+        [ ! -d $TREE ] && echo -e "\n$ERROR_LOG\n==> Teknoparrot is not installed." && exit
         variableTree
-        if [[ ! -d $TREE ]]; then
-            echo -e "\n[$REDCOLOR ERROR $NC]\n==> Teknoparrot is not installed."
-        else
-            fileExistenceChecker
-            echo -e "\n[$GREENCOLOR DONE $NC]\n==> Teknoparrot successfully removed!"
-            sleep 2
-        fi
+        [ ! -d $TREE ] && echo -e "\n$DONE_LOG\n==> Teknoparrot successfully removed!"
         exit
     ;;
 esac
@@ -145,7 +136,6 @@ esac
 clear
 creationTree
 variableTree
-fileExistenceChecker
 runner
 dependencyInstall
 executableCreation
